Moved BGP related code to bgp.py
